requirements.txt
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
ffmpeg==1.4
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.9.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0